test: migrate 3 ESLint RuleTester files to bun:test (3 left)
ESLint's RuleTester registers describe/it blocks dynamically when
`.run(...)` is called. Bun:test does not allow new test()
registrations from inside a running test() callback (vitest does),
so the previous structure -- describe(...) wrapping test() callbacks
that called ruleTester.run() -- failed under bun:test.

Restructure: move ruleTester.run() to module top-level, with all
valid + invalid rule cases consolidated into one call per rule.
RuleTester then registers its own describe/it tree at module load
(when bun:test does accept registrations). Bind
RuleTester.describe/it to bun:test's exports so the registration
hooks resolve.

Migrated:
- packages/eslint-plugin-routecraft/test/batch-before-from.bun.test.ts
  (9 tests)
- packages/eslint-plugin-routecraft/test/require-named-route.bun.test.ts
  (15 tests)
- packages/eslint-plugin-routecraft/test/single-to-per-route.bun.test.ts
  (8 tests)

Local: 1057 bun + 60 vitest = 1117 tests pass.
